refactor: split integration env workspace helpers
