+++ .sampo/changesets/1004-split-integration-env-runtime.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split integration environment workspace add helpers into focused runtime modules without changing generated output.
